VERY UNSTABLE AND BUGGY COMMIT - DO NOT USE RECREATIONALLY
------------------------------------------

Added the ability for a user to tap on their recent creations and get back to editing them. I am yet to add a 'Save edits' feature, so keep that in mind.

I've also added a RecentCreationsListener.kt.

There have been major code modifications in many files.

The code as it stands now is very messy and bug prone! Bugs and messy code will be fixed in future commits.

Please try and find a more stable commit if you are wanting to use the app.