MAJOR UPDATE/CHANGE
Prior to this commit, the user had to press on each individual pixel, in this commit I added a dragging functionality - which makes the user's life much more easier than having to manually tap on each pixel.

Depending on the grid size it may not work as expected, so keep that fact into consideration. You may also encounter bugs. The tool will be refined in future commits.