MAJOR FEATURE ADDITION:
Now, whenever the user wants to create a new canvas by pressing the floating action button in activity_main.xml, they will be led to a Fragment instead of a dialog. This is because whenever a fragment is used, there is more space to add more options as to how the user will customise their new canvas (as in background colour, et cetera).