Added two methods to CanvasActivityTest.kt:
The first method titled 'goToFindAndReplaceFragment' will tap on the 'Find and Replace' Button, which will result in the navigation to the FindAndReplaceFragment.

The second method titled 'goToColorPickerFragment' will long press on either the primary or secondary color View, which will result in the navigation to the ColorPickerFragment.

Both of these functions are utility functions which will be used in unit tests in CanvasActivityTest.kt to minimise the lines of code.

Method code will be added in future commits.