EARLY UNSTABLE EXTREMELY LARGE MAJOR COMMIT AND APP RE-MODIFICATION AND CHANGE.
Note: do NOT clone or branch this application from this commit as it is NOT stable.

Major changes have been made in PyxlMoose. In this commit, the user (should, although it should be known that there are many bugs) can save their recent creations persistently.

This means, after the user creates a pixel art creation, closes the app, and reopens it, it will still remain. The user can also update their pixel art creation or even delete it.

Prior to this commit, data was not stored persistently. This meant that once the user created something, left the application, and then came back - their data was gone. This was one of the key problems PyxlMoose faced and one it needed to address if it wanted to become a usable app.

I am NOT going to be giving an explanation about what features have been added in terms of code, anyone wants to understand how this database was implemented is free to do so on their own, as explaining all of the classes and code re-modifications can take hundreds of lines.

Again, this was only committed so I do not lose my progress - this is NOWHERE near stable. Do not clone or fork the app from here!