EARLY UNSTABLE EXTREMELY LARGE MAJOR COMMIT AND APP RE-MODIFICATION A… …ND CHANGE.
Note: do NOT clone or branch this application from this commit as it is NOT stable.

In this commit, I have refactored the handling of the pixel art canvas from a `RecyclerView` to a `Canvas` because (as mentioned in the Project):

a. Prior to this commit, PyxlMoose did not support canvas sizes of over 100 by 100. Transitioning to a Canvas now allows the user to create a larger canvas with good performance.

b. Transitioning pixel art canvas code to a `Canvas` will allow PyxlMoose will allow the addition of more advanced features in the future.

c. Using a RecyclerView to represent pixels is slow and inefficient, and there have also been many rendering issues using a RecyclerView. With `Canvas` these rendering issues are gone.

Again, this commit does contain a lot of bugs and the code IS NOT clean in any way!