Added the ability to extract projects to their raw dimensions (the user also has the options of extracting to scaled dimensions as well). This closes issue #184 (fix #184).
More changes are needed to be done in future commits to stabilize this feature and improve the user interface.

'Save as' has been transitioned into a single menu item titled 'Export'.

When the user clicks 'Export', a dialog is shown from which the user can input values such as the file's title, the file's type, and the file's resolution (raw or scaled).