Begin work on new caching/roi system for graphics backend.
New `GBitmap.overlay` which works much better than old one.

Fix some of the Flutter code to work better with the Provider.

Layers kind of work now in the Flutter app but more work needs to be done.

A bit of a messy commit as lots of features in progress. Mostly just changes in `graphics` and `app` package.